drivers: adc-dac: ad74416h: update dac code type
The dac code value is 16-bit wide. Use the appropriate data type.

+++ drivers/adc-dac/ad74416h/ad74416h.h
@@ -636,7 +636,7 @@ struct ad74416h_desc {
 
 /** Converts a millivolt value in the corresponding DAC 13 bit code */
 int ad74416h_dac_voltage_to_code(struct ad74416h_desc *, int32_t,
-				 uint32_t *, uint32_t);
+				 uint16_t *, uint32_t);
 
 /** Converts a microamp value in the corresponding DAC 16 bit code */
 int ad74416h_dac_current_to_code(uint32_t, uint16_t *);
